requirements.txt
absl-py==2.1.0
keras
matplotlib==3.9.2
numpy
tensorflow # Adicionando o TensorFlow diretamente
tqdm==4.67.1
pillow>=8.0.0
keras-preprocessing
